Rename choropleth.py to map_plotter.py
This seems to reflect a lot of what the class does, and there may be a
  case where we don't want a choropleth map.

Plus this means that I don't have to look up how to spell it as often!
