+++ src/scout_map.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import numpy as np
-from src.choropleth import ChoroplethMapPlotter
+from src.map_plotter import ChoroplethMapPlotter
 from src.census_data import CensusData
 from src.census_merge_postcode import CensusMergePostcode
 import src.log_util as log_util
